% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_new.R
\name{NR_new}
\alias{NR_new}
\title{Netwon Raphson}
\usage{
NR_new(
  delta,
  z,
  time,
  knot = 10,
  facility = NULL,
  M_stop = 10000,
  tol = 10^(-6),
  rate = 0.001
)
}
\arguments{
\item{delta}{event indicator.}

\item{z}{Covariate matrix.}

\item{time}{observed event time.}

\item{knot}{number of basis functions for time-varing effects.}

\item{facility}{strata.}

\item{M_stop}{maximum stopping iterations.}

\item{tol}{the convergence threshlod.}

\item{rate}{}
}
\value{
return a list of the following values:
\itemize{
  \item theta:
  \item test_SGD_all:
  \item b_spline:
}
}
\description{
Netwon Raphson
}
\examples{

# the SGD
# generate the simuluation data
data2 <- simul(N = 2000, N_Strata = 5, p=5 )
result1 <- NR_new(data2$delta, data2$z, data2$time, knot = 10 ,
                  facility = data2$facility, M_stop = 10000, tol= 10^(-6),
                  rate=0.001)

}
