% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR_new.R
\name{NR_new}
\alias{NR_new}
\title{Netwon Raphson}
\usage{
NR_new(
  delta,
  z,
  time,
  knot = 10,
  facility = NULL,
  M_stop = 10000,
  tol = 10^(-6),
  rate = 0.001
)
}
\arguments{
\item{delta}{event indicator.}

\item{z}{Covariate matrix.}

\item{time}{observed event time.}

\item{knot}{number of basis functions for time-varing effects.}

\item{facility}{strata.}

\item{M_stop}{maximum stopping iterations.}

\item{tol}{the convergence threshlod.}

\item{rate}{}
}
\value{
return a list of the following values
*theta : 
*test_SGD_all :
*b_spline :
}
\description{
Netwon Raphson
}
\examples{

}
